feat(email): implement email service module and templates
- Create new MailerModule and MailerService in libs/integrations
- Add email templates for verification and welcome emails
- Refactor EmailService to use MailerService
- Update auth module to use EmailModule instead of EmailService
