+++ libs/integrations/email/email.module.ts
@@ -1,0 +1,11 @@
+import { Module } from '@nestjs/common';
+import { ConfigModule } from '@nestjs/config';
+import { MailerService } from './mailer.service';
+
+@Module({
+  imports: [ConfigModule],
+  providers: [MailerService],
+  exports: [MailerService],
+})
+export class MailerModule {}
+
