+++ libs/integrations/email/index.ts
@@ -1,0 +1,3 @@
+export * from './email.module';
+export * from './mailer.service';
+
