+++ libs/integrations/index.ts
@@ -1,3 +1,4 @@
 // Central exports for integrations
 export * from './aliyun';
 export * from './lark';
+export * from './email';
